Add three base daemon builders to enable more customization
