+++ src/main/java/com/liveramp/daemon_lib/builders/BaseBlockingDaemonBuilder.java
@@ -1,0 +1,44 @@
+package com.liveramp.daemon_lib.builders;
+
+import java.io.IOException;
+
+import org.jetbrains.annotations.NotNull;
+
+import com.liveramp.daemon_lib.JobletCallback;
+import com.liveramp.daemon_lib.JobletConfig;
+import com.liveramp.daemon_lib.JobletConfigProducer;
+import com.liveramp.daemon_lib.JobletFactory;
+import com.liveramp.daemon_lib.executors.JobletExecutor;
+import com.liveramp.daemon_lib.executors.JobletExecutors;
+
+public abstract class BaseBlockingDaemonBuilder<T extends JobletConfig, E extends BaseBlockingDaemonBuilder<T, E>> extends BaseDaemonBuilder<T, E> {
+
+  private final JobletFactory<T> jobletFactory;
+  private JobletCallback<T> successCallback;
+  private JobletCallback<T> failureCallback;
+
+  protected BaseBlockingDaemonBuilder(String identifier, JobletFactory<T> jobletFactory, JobletConfigProducer<T> configProducer) {
+    super(identifier, configProducer);
+    this.jobletFactory = jobletFactory;
+
+    this.successCallback = new JobletCallback.None<>();
+    this.failureCallback = new JobletCallback.None<>();
+  }
+
+  public E setSuccessCallback(JobletCallback<T> callback) {
+    this.successCallback = callback;
+    return self();
+  }
+
+  public E setFailureCallback(JobletCallback<T> callback) {
+    this.failureCallback = callback;
+    return self();
+  }
+
+  @NotNull
+  @Override
+  protected JobletExecutor<T> getExecutor() throws IllegalAccessException, IOException, InstantiationException {
+    return JobletExecutors.Blocking.get(jobletFactory, successCallback, failureCallback);
+  }
+
+}
