+++ src/main/java/com/liveramp/daemon_lib/builders/BaseForkingDaemonBuilder.java
@@ -1,0 +1,70 @@
+package com.liveramp.daemon_lib.builders;
+
+import java.io.File;
+import java.io.IOException;
+import java.util.Map;
+
+import com.google.common.collect.Maps;
+import org.jetbrains.annotations.NotNull;
+
+import com.liveramp.daemon_lib.JobletCallback;
+import com.liveramp.daemon_lib.JobletConfig;
+import com.liveramp.daemon_lib.JobletConfigProducer;
+import com.liveramp.daemon_lib.JobletFactory;
+import com.liveramp.daemon_lib.executors.JobletExecutor;
+import com.liveramp.daemon_lib.executors.JobletExecutors;
+import com.liveramp.daemon_lib.executors.forking.ProcessJobletRunner;
+
+public abstract class BaseForkingDaemonBuilder<T extends JobletConfig, E extends BaseForkingDaemonBuilder<T, E>> extends BaseDaemonBuilder<T, E> {
+
+  private final String workingDir;
+  private final Class<? extends JobletFactory<T>> jobletFactoryClass;
+  private int maxProcesses;
+  private Map<String, String> envVariables;
+  private JobletCallback<? super T> successCallback;
+  private JobletCallback<? super T> failureCallback;
+  private ProcessJobletRunner jobletRunner;
+
+  private static final int DEFAULT_MAX_PROCESSES = 1;
+  private static final Map<String, String> DEFAULT_ENV_VARS = Maps.newHashMap();
+
+  protected BaseForkingDaemonBuilder(String workingDir, String identifier, Class<? extends JobletFactory<T>> jobletFactoryClass, JobletConfigProducer<T> configProducer, ProcessJobletRunner jobletRunner) {
+    super(identifier, configProducer);
+    this.workingDir = workingDir;
+    this.jobletFactoryClass = jobletFactoryClass;
+    this.jobletRunner = jobletRunner;
+
+    maxProcesses = DEFAULT_MAX_PROCESSES;
+    envVariables = DEFAULT_ENV_VARS;
+    successCallback = new JobletCallback.None<>();
+    failureCallback = new JobletCallback.None<>();
+  }
+
+  public E setMaxProcesses(int maxProcesses) {
+    this.maxProcesses = maxProcesses;
+    return self();
+  }
+
+  public E addToEnvironmentVariables(String envVar, String value) {
+    this.envVariables.put(envVar, value);
+    return self();
+  }
+
+  public E setSuccessCallback(JobletCallback<? super T> callback) {
+    this.successCallback = callback;
+    return self();
+  }
+
+  public E setFailureCallback(JobletCallback<? super T> callback) {
+    this.failureCallback = callback;
+    return self();
+  }
+
+  @NotNull
+  @Override
+  protected JobletExecutor<T> getExecutor() throws IllegalAccessException, IOException, InstantiationException {
+    final String tmpPath = new File(workingDir, identifier).getPath();
+    return JobletExecutors.Forked.get(notifier, tmpPath, maxProcesses, jobletFactoryClass, envVariables, successCallback, failureCallback, jobletRunner);
+  }
+
+}
