+++ src/main/java/com/liveramp/daemon_lib/builders/ThreadingDaemonBuilder.java
@@ -1,54 +1,16 @@
 package com.liveramp.daemon_lib.builders;
 
-import java.io.IOException;
-
-import org.jetbrains.annotations.NotNull;
-
-import com.liveramp.daemon_lib.JobletCallback;
 import com.liveramp.daemon_lib.JobletConfig;
 import com.liveramp.daemon_lib.JobletConfigProducer;
 import com.liveramp.daemon_lib.JobletFactory;
-import com.liveramp.daemon_lib.executors.JobletExecutor;
-import com.liveramp.daemon_lib.executors.JobletExecutors;
 
-public class ThreadingDaemonBuilder<T extends JobletConfig> extends BaseDaemonBuilder<T, ThreadingDaemonBuilder<T>> {
-
-  private final JobletFactory<T> jobletFactory;
-  private JobletCallback<T> successCallback;
-  private JobletCallback<T> failureCallback;
-  private int maxThreads;
-
-  private static final int DEFAULT_MAX_THREADS = 1;
-
-
+public class ThreadingDaemonBuilder<T extends JobletConfig> extends BaseThreadingDaemonBuilder<T, ThreadingDaemonBuilder<T>> {
   public ThreadingDaemonBuilder(String identifier, JobletFactory<T> jobletFactory, JobletConfigProducer<T> configProducer) {
-    super(identifier, configProducer);
-    this.jobletFactory = jobletFactory;
-
-    this.maxThreads = DEFAULT_MAX_THREADS;
-
-    this.successCallback = new JobletCallback.None<>();
-    this.failureCallback = new JobletCallback.None<>();
+    super(identifier, jobletFactory, configProducer);
   }
 
-  public ThreadingDaemonBuilder<T> setMaxThreads(int maxThreads) {
-    this.maxThreads = maxThreads;
-    return this;
-  }
-
-  public ThreadingDaemonBuilder<T> setSuccessCallback(JobletCallback<T> successCallback) {
-    this.successCallback = successCallback;
-    return this;
-  }
-
-  public ThreadingDaemonBuilder<T> setFailureCallback(JobletCallback<T> failureCallback) {
-    this.failureCallback = failureCallback;
-    return this;
-  }
-
-  @NotNull
   @Override
-  protected JobletExecutor<T> getExecutor() throws IllegalAccessException, IOException, InstantiationException {
-    return JobletExecutors.Threaded.get(maxThreads, jobletFactory, successCallback, failureCallback);
+  protected ThreadingDaemonBuilder<T> self() {
+    return this;
   }
 }
